pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bruno-mcp"
version = "0.2.0"
description = "MCP server for Bruno API collections"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Jack Mulligan"}]
readme = "README.md"
keywords = ["mcp", "bruno", "api"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "fastmcp>=0.4.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "build>=1.4.0",
    "twine>=6.2.0",
    "debugpy>=1.8.20",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[project.scripts]
bruno-mcp = "bruno_mcp.__main__:main"

[project.urls]
Repository = "https://github.com/jackmulligan-ire/bruno-mcp"
